added component folder, put components inside it
